Assistant checkpoint
Assistant generated file changes:
- client/src/components/layout/sidebar.tsx: Add login button to sidebar
Add login button to sidebar
Add login/logout functionality

---

User prompt:

this isn't working. I want to test if it's because I am not logged in or not.  There isn't even a log in page.  Create a button that says "Log in / Create Account" and allow me to create our first account